refactor(llm): split /graph/extract into a dedicated API surface
Separate graph extraction from the RAG API and harden its
request-scoped graph config contract.

- move GraphExtractRequest/GraphExtractClientConfig to
  api/models/graph_extract_requests.py and GraphExtractResponse to
  api/models/graph_extract_responses.py
- move the route to api/graph_extract_api.py (graph_extract_http_api);
  register it from demo/rag_demo/app.py and drop graph_api.py
- return a stable GraphExtractResponse envelope (status/result/
  warnings/meta) instead of the raw flow dict
- make the endpoint property-graph only (remove "triples" from the enum)
- add a dedicated GraphExtractClientConfig (extra="forbid", all-optional,
  no url field) so 127.0.0.1:8080/hugegraph defaults can't leak in
- force the connection url to the configured server url; reject
  request-supplied urls
- keep strict checks: named schema requires client_config,
  client_config.graph must match the schema name, inline schema +
  client_config rejected, no field-by-field fallback into globals
- rewrite tests as test_graph_extract_api.py covering the public contract
