+++ hugegraph-llm/src/hugegraph_llm/flows/graph_extract.py
@@ -17,6 +17,7 @@ import json
 
 from pycgraph import GPipeline
 
+from hugegraph_llm.config import huge_settings
 from hugegraph_llm.flows.common import BaseFlow
 from hugegraph_llm.nodes.document_node.chunk_split import ChunkSplitNode
 from hugegraph_llm.nodes.hugegraph_node.schema import SchemaNode
@@ -49,8 +50,9 @@ class GraphExtractFlow(BaseFlow):
         prepared_input.extract_type = extract_type
         client_config = kwargs.get("client_config")
         if client_config:
+            # URL stays server-controlled; only identity/graphspace are request-scoped.
             prepared_input.graph_client_config = {
-                "url": client_config.url,
+                "url": huge_settings.graph_url,
                 "user": client_config.user,
                 "pwd": client_config.pwd,
                 "graphspace": client_config.gs,
